# docs/swagger.yaml
basePath: /api/v1
definitions:
  favorite.AddRequest:
    properties:
      entity_id:
        type: string
      entity_type:
        enum:
        - casting
        - profile
        type: string
    required:
    - entity_id
    - entity_type
    type: object
  github_com_mwork_mwork-api_internal_pkg_response.ErrorInfo:
    properties:
      code:
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  github_com_mwork_mwork-api_internal_pkg_response.Meta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      pages:
        type: integer
      total:
        type: integer
    type: object
  github_com_mwork_mwork-api_internal_pkg_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.ErrorInfo'
      meta:
        $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Meta'
      success:
        type: boolean
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
  description: API server for model agency aggregator.
  termsOfService: http://swagger.io/terms/
  title: MWork API
  version: "1.0"
paths:
...
  /favorites:
    get:
      parameters:
      - description: Тип фильтра (casting|profile)
        in: query
        name: type
        type: string
      - description: Лимит
        in: query
        name: limit
        type: integer
      - description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Список избранного
      tags:
      - Favorite
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные избранного
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/favorite.AddRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Добавить в избранное
      tags:
      - Favorite
  /favorites/{type}/{id}:
    delete:
      parameters:
      - description: Тип сущности (casting|profile)
        in: path
        name: type
        required: true
        type: string
      - description: ID сущности
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Удалить из избранного
      tags:
      - Favorite
  /favorites/{type}/{id}/check:
    get:
      parameters:
      - description: Тип сущности (casting|profile)
        in: path
        name: type
        required: true
        type: string
      - description: ID сущности
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: boolean
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_mwork_mwork-api_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Проверить наличие в избранном
      tags:
      - Favorite
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"